Ensure HTTP client reads the entire body
According to https://godoc.org/net/http#Client.Do, the http.Transport
may be unable to reuse the connection if not EOF is returned from Body,
and Close is not called on it. Assuming the JSON decoder will not read
more than it has to, this will never happen. This commit ensures that we
trigger EOF by copying the remaining bytes (whitespace, if any) to
/dev/null before returning the result.